Move constants.ts to inside "tests" folder.
