+++ tests/unauthorized/credential-verify/negative.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 import { StatusCodes } from 'http-status-codes';
-import { INVALID_CREDENTIAL, INVALID_JWT_TOKEN, NOT_MATCHED_CREDENTIAL_AND_JWT } from '../constants';
+import { INVALID_CREDENTIAL, INVALID_JWT_TOKEN, NOT_MATCHED_CREDENTIAL_AND_JWT } from '../../constants';
 
 test('[Negative] It cannot verify credential with an invalid credential body where credential and JWT are different', async ({ request }) => {
     const response = await request.post('/credential/verify', {
